reworked windows but kinda not really...
- Forgot how QWidget system worked so redid the whole parent reference
system of Cuter classes to make “Screens,” only to realize the Window
class literally accomplished that…so un-did all of it, and made Window
a Screen class.
- All objects now…well, they do what they did before, they just take in
the name of the screen object they’re referencing which is stored in a
dict
- Should help with adding some more complicated functionality, like
grade inputting, because hoo boy that main screen was getting cramped.